Fix imports for moved chat service
