+++ src/app/feature/chat/chat-service/offline/contact-field.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { LocalStorageService } from '../../local-storage/local-storage.service';
+import { LocalStorageService } from 'src/app/shared/services/local-storage/local-storage.service';
 
 @Injectable({
   providedIn: 'root'
